pyproject.toml
[project]
name = "dagster-datacontract"
version = "0.4.0"
description = "Load metadata and asset check spesifications from data contracts."
authors = [
    { name = "Fredrik Bakken", email = "fredrik@dataheim.io" }
]
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "dagster>=1.10.10",
    "datacontract-cli>=0.10.23",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "dagster-webserver>=1.10.10",
    "polars>=1.27.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
extend-select = [
    "UP", # pyupgrade
    "I",  # re-order imports in alphabetic order
]
